created selectors + some refactoring
