+++ src/api/api.ts
@@ -1,10 +1,11 @@
 import axios, {AxiosInstance} from 'axios';
 import {RequestInterceptor, ResponseInterceptor} from './interceptors.ts';
+import {BACKEND_URL, REQUEST_TIMEOUT} from '../const.ts';
 
 export const createAPI = (): AxiosInstance => {
   const axiousInstance = axios.create({
-    baseURL: 'https://13.design.pages.academy/wtw',
-    timeout: 5000,
+    baseURL: BACKEND_URL,
+    timeout: REQUEST_TIMEOUT,
   });
 
   axiousInstance.interceptors.request.use(RequestInterceptor);
