+++ src/api/utils.ts
@@ -1,7 +1,9 @@
+import {TOKEN_KEY} from '../const.ts';
+
 export function setToken(token: string) {
-  localStorage.setItem('token', token);
+  localStorage.setItem(TOKEN_KEY, token);
 }
 
 export function getToken() {
-  return localStorage.getItem('token') ?? '';
+  return localStorage.getItem(TOKEN_KEY) ?? '';
 }
