+++ src/components/add-review-form/add-review-form.tsx
@@ -2,8 +2,8 @@ import {FormEvent, useEffect, useState} from 'react';
 import {postComment} from '../../store/thunks.ts';
 import {useAppDispatch} from '../../hooks/redux-typed-hooks.ts';
 import {Navigate, useNavigate, useParams} from 'react-router-dom';
-import {AppRoute} from '../../const.ts';
 import React from 'react';
+import {AppRoute} from '../../types/enums.ts';
 
 
 function AddReviewForm(): JSX.Element {
