+++ src/components/film-card/film-card.tsx
@@ -1,8 +1,8 @@
 import {HTMLAttributes} from 'react';
-import {AppRoute} from '../../const';
 import {Link} from 'react-router-dom';
 import VideoPlayer from '../video-player/video-player';
 import {FilmShallow} from '../../types/film-shallow.ts';
+import {AppRoute} from '../../types/enums.ts';
 
 interface ArticleProps extends HTMLAttributes<HTMLDivElement> {
   film: FilmShallow;
