+++ src/components/login-form/login-form.tsx
@@ -2,8 +2,8 @@ import {ChangeEvent, FormEvent, useEffect, useState} from 'react';
 import {postLoginData} from '../../store/thunks.ts';
 import {useAppDispatch, useAppSelector} from '../../hooks/redux-typed-hooks.ts';
 import {useNavigate} from 'react-router-dom';
-import {AppRoute} from '../../const.ts';
 import {selectIsAuthenticated} from '../../store/user/user-slice.selectors.ts';
+import {AppRoute} from '../../types/enums.ts';
 
 export const LoginForm = () => {
   const dispatch = useAppDispatch();
