+++ src/hooks/use-film.ts
@@ -2,8 +2,8 @@ import {useAppDispatch, useAppSelector} from './redux-typed-hooks.ts';
 import {useNavigate, useParams} from 'react-router-dom';
 import {useEffect} from 'react';
 import {getFilmById} from '../store/thunks.ts';
-import {AppRoute} from '../const.ts';
 import {selectFilm, selectIsFilmLoading} from '../store/film/film-slice.selectors.ts';
+import {AppRoute} from '../types/enums.ts';
 
 export const useFilm = () => {
   const film = useAppSelector(selectFilm);
