+++ src/pages/player-page/player-page.test.tsx
@@ -46,9 +46,7 @@ describe('Component: Player page', () => {
 
     HTMLVideoElement.prototype.play = vi.fn().mockReturnValue({
       then: () => ({
-        catch: () => {
-          // do nothing
-        }
+        catch: () => undefined
       }),
     });
     HTMLVideoElement.prototype.pause = vi.fn();
