+++ src/store/main/main-slice.ts
@@ -2,12 +2,12 @@ import {createSlice, PayloadAction} from '@reduxjs/toolkit';
 
 import {MainState} from '../../types/state.ts';
 import {getAllFilms, getPromoFilm} from '../thunks.ts';
-import {DefaultFilmGenre} from '../../const.ts';
+import {DEFAULT_FILM_GENRE} from '../../const.ts';
 
 const initialState: MainState = {
   films: [],
   isFilmListLoading: true,
-  genre: DefaultFilmGenre,
+  genre: DEFAULT_FILM_GENRE,
   promoFilm: null,
 };
