+++ src/types/common-interfaces.ts
@@ -1,0 +1,5 @@
+import {Film} from './film.ts';
+
+export interface CommonDescriptionProps {
+  film: Film;
+}
